chore: add SettingService test
