+++ src/ingress-nginx-validate-jwt/Controllers/AuthController.cs
@@ -3,6 +3,7 @@ using System.Net.Http.Headers;
 using System.Text.Json;
 using System.Web;
 using ingress_nginx_validate_jwt.Constants;
+using ingress_nginx_validate_jwt.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Primitives;
 using Microsoft.IdentityModel.Tokens;
