+++ src/ingress-nginx-validate-jwt/Services/ISettingsService.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.IdentityModel.Protocols.OpenIdConnect;
+
+namespace ingress_nginx_validate_jwt.Services;
+
+public interface ISettingsService
+{
+    Task<OpenIdConnectConfiguration> GetConfiguration(CancellationToken cancellationToken = default);
+}
